+++ src/containerapp/azext_containerapp/_help.py
@@ -407,6 +407,11 @@ helps['containerapp service mariadb'] = """
     short-summary: Commands to manage the mariadb service for the Container Apps environment.
 """
 
+helps['containerapp service qdrantvdb'] = """
+    type: group
+    short-summary: Commands to manage the qdrantvdb service for the Container Apps environment.
+"""
+
 helps['containerapp service redis create'] = """
     type: command
     short-summary: Command to create the redis service.
@@ -427,6 +432,11 @@ helps['containerapp service mariadb create'] = """
     short-summary: Command to create the mariadb service.
 """
 
+helps['containerapp service qdrantvdb create'] = """
+    type: command
+    short-summary: Command to create the qdrantvdb service.
+"""
+
 helps['containerapp service redis delete'] = """
     type: command
     short-summary: Command to delete the redis service.
@@ -445,6 +455,11 @@ helps['containerapp service kafka delete'] = """
 helps['containerapp service mariadb delete'] = """
     type: command
     short-summary: Command to delete the mariadb service.
+"""
+
+helps['containerapp service qdrantvdb delete'] = """
+    type: command
+    short-summary: Command to delete the qdrantvdb service.
 """
 
 helps['containerapp env update'] = """